requirements.txt
requests==2.7.0
semantic-version==2.4.2
python-dateutil==2.6.1
